requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-pinecone
pinecone-client
pypdf
python-dotenv
gunicorn
sentence-transformers